Add delete app method to helpers
