+++ app/controllers/project.py
@@ -2,7 +2,7 @@ from app.helpers.prometheus import prometheus
 from app.helpers.alias import create_alias
 from app.helpers.admin import is_owner_or_admin, is_current_or_admin
 from app.helpers.role_search import has_role
-from app.helpers.kube import create_kube_clients
+from app.helpers.kube import create_kube_clients, delete_cluster_app
 from app.models.user import User
 from app.models.clusters import Cluster
 from app.models.project import Project
@@ -13,7 +13,6 @@ import json
 from flask_restful import Resource, request
 from kubernetes import client
 from flask_jwt_extended import jwt_required, get_jwt_identity, get_jwt_claims
-from app.controllers.app import delete_cluster_app
 
 class ProjectsView(Resource):
